+++ src/map/tests.rs
@@ -1639,20 +1639,27 @@ fn test_reopen_mmap() {
 #[cfg(feature = "memmap")]
 #[cfg_attr(miri, ignore)]
 fn test_reopen_mmap2() {
+  use rand::seq::SliceRandom;
+
   let dir = tempfile::tempdir().unwrap();
   let p = dir.path().join("reopen_skipmap2");
   {
     let l = SkipMap::mmap_mut_with_comparator(&p, ARENA_SIZE, true, Ascend).unwrap();
-    for i in 0..1000 {
+    let mut data = (0..1000).collect::<Vec<usize>>();
+    data.shuffle(&mut rand::thread_rng());
+    for i in data {
       l.insert(i as u64, &key(i), &new_value(i)).unwrap();
     }
     l.flush_async().unwrap();
     assert_eq!(l.max_version(), 999);
+    assert_eq!(l.min_version(), 0);
   }
 
   let l = SkipMap::<u64, Ascend>::mmap_with_comparator(&p, false, Ascend).unwrap();
   assert_eq!(1000, l.len());
-  for i in 0..1000 {
+  let mut data = (0..1000).collect::<Vec<usize>>();
+  data.shuffle(&mut rand::thread_rng());
+  for i in data {
     let k = key(i);
     let ent = l.get(i as u64, &k).unwrap();
     assert_eq!(new_value(i), ent.value());
@@ -1660,6 +1667,7 @@ fn test_reopen_mmap2() {
     assert_eq!(ent.key(), k);
   }
   assert_eq!(l.max_version(), 999);
+  assert_eq!(l.min_version(), 0);
 }
 
 struct Person {